nothing

+++ SickVideosProject/SickVideos/src/main/java/com/sickboots/sickvideos/billing/IabHelper.java
@@ -778,7 +778,7 @@ public class IabHelper {
       logDebug("Bundle with null response code, assuming OK (known issue)");
       return BILLING_RESPONSE_RESULT_OK;
     } else if (o instanceof Integer)
-      return ((Integer) o).intValue();
+      return (Integer) o;
     else if (o instanceof Long)
       return (int) ((Long) o).longValue();
     else {
